:heavy_plus_sign: Added Three.js back under testIndex.html
# Three.js is coming back! :100: :100: :100:
## .gitignore:
- css files were copied for local testing, this ignores them

## o.png and x.png were moved here from the asa-dillahunty.github.io repo

## interface-noThree.js.js was ripped apart into the files:

### botLogic.js
contains logic for creating bots including the methods
- getMove
- canWin
- getScore
- newBot

### gameLogic.js
Contains the logic for the actual games and interface modifying functions. Important methods/variables:
- gameBoard
- turn
- howToPlay
- boardInit
- makeClickable/makeNotClickable
- clickedSquare
- sendMove
- startGame
- showModal/hideModal

### onlineLogic.js
Contains logic for online connections and games. This file is soon to undergo a massive change as this logic sucks :-1:
- onlineGameData
- dataStreamHandler
- onlineHandler

### threejsLogic.js
This file contains all the logic necessary to render the Threejs model (aside from the actual source files lol). This file ***should*** be completely optional.
- threejsInit
- clearThreejsBoard
- markThreejsBoard

## testIndex.html
Should almost be an exact copy of index.html besides the javascript files linked at the bottom of the page

## qubic.css
moved some javascript logic for adding styling to here

## Threejs source files
moved and updated the Threejs source files